nfp: sgw: add ipsec cmsg api
IPSec is needed to be supported in sgw kernel driver. sgw ipsec use
cmsg to send configuration message, we reuse original
nfp_ipsec_cfg_add_sa api, but some modifications are necessary for
compile. NFP_FLOWER_CMSG_TYPE_IPSEC is 27 as 26 is already used in PMD
driver, kernel driver and PMD driver use same firmware. The fields in
the cmsg also needs to be byte swapped to be used on the sgw firmware.

Signed-off-by: Sifei Ni <sifei.ni@corigine.com>
Signed-off-by: Xuehao Yin <xuehao.yin@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>