workflows: build: Set up and run in own docker
Remove github's docker engine and replace it by setting up a
docker environment that is run in the background. Stop and remove
the docker environment at the end of the workflow.

Run the commands in the created container. In order for this to work
the Ubuntu job needed to be switched to use docker as well, since the
local build action now uses the docker command as it's shell, which
breaks not dockerized use.

Any Node16 issues are solved by running commands in the
container.

Co-authored-by: Annemie Potgieter <annemie.potgieter@corigine.com>
Signed-off-by: Louis Peens <louis.peens@corigine.com>