src: crdma: crdma_ucif: Remove redundant const qualifier
Remove instances where a redundant "const" qualifier is included in the
function definition.

The functions are declared as:
"const char * const func_name()"

The first "const" instance indicated that the function returns a pointer
to a constant char, in other words, that the data being pointed to
cannot be changed through the returned pointer.

The second "const" indicateds that the pointer itself is constant and
cannot be reassigned to point to a different memory address. This is
redundant because it is being applied to a temporary value, and since
you cannot reassign a temporary value, the qualifier has no practical
effect.

This redundancy was exposed when attempting to build the nfp kernel
driver on the CentOS Stream 10 operating system, which ships with a
newer compiler version (gcc 14.2) than on CentOS Stream 9 (gcc 11.5).

Signed-off-by: James Hershaw <james.hershaw@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>