nfp: diversify the ways of getting serial number
Currently we combine vendor, partno and serial to make a
serial number of card, which is used in devlink dev info
and vpd attribute of sys file system. In some scenarios,
serial number is not generated as this, nfp driver can
get directly from hwinfo. This commit diversifies the ways
of getting serial number, getting directly is preferred.

Signed-off-by: Guibin Xia <guibin.xia@corigine.com>
Signed-off-by: Kunlun Mao <kunlun.mao@corigine.com>
Reviewed-by: Louis Peens <louis.peens@corigine.com>